Restructured `opt` module.
